+++ subscription.php
@@ -6,9 +6,9 @@
  * glFusion configuration system, such as allowed file types.
  *
  * @author      Lee Garner <lee@leegarner.com>
- * @copyright   Copyright (c) 2010-2019 Lee Garner
+ * @copyright   Copyright (c) 2010-2020 Lee Garner
  * @package     subscription
- * @version     v1.0.0
+ * @version     v1.0.1
  * @license     http://opensource.org/licenses/gpl-2.0.php
  *              GNU Public License v2 or later
  * @filesource
@@ -23,7 +23,7 @@ global $_CONF_SUBSCR;
 
 $_CONF_SUBSCR['pi_name']            = 'subscription';
 $_CONF_SUBSCR['pi_display_name']    = 'Subscriptions';
-$_CONF_SUBSCR['pi_version']         = '1.0.0';
+$_CONF_SUBSCR['pi_version']         = '1.0.1';
 $_CONF_SUBSCR['gl_version']         = '1.7.0';
 $_CONF_SUBSCR['pi_url']             = 'http://www.leegarner.com';